// src/core/types/entities/iron-portal/dashboard/pbi-equipment/component-replacement/ListItem.ts
export type ListItem = {
    componentReplacementId: number
    site: string
    model: string
    equipment: string
    component: string
    componentAssignmentId: number
    replacementDate: string
    replacementSmu: number
    hmoffset: number
    frameHours: number
    transactionCount: number
    isActive: boolean
    createdBy: string
    createdOn: string
    changedBy: string
    changedOn: string
}