// src/core/types/entities/iron-portal/dashboard/pbi-equipment/component-replacement/FilterData.ts
export type FilterData = {
    site: string
    siteTo: string
    model: string
    modelTo: string
    equipment: string
    equipmentTo: string
    component: string
    componentTo: string
    replacementDate: string
    replacementDateTo: string
    replacementSmu: string
    replacementSmuTo: string
    frameHours: string
    frameHoursTo: string
    page: number
    pageSize: number
    order: string
    ver: string
}